azure-function-app/PgTriggerFunction/requirements.txt
azure-functions
psycopg2-binary>=2.9.9
mlflow>=2.10.0
transformers>=4.37.0
torch>=2.2.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
